Starts process for: Implement Environment Mods
