+++ common/src/main/java/org/kakara/core/common/EnvTypeable.java
@@ -1,0 +1,8 @@
+package org.kakara.core.common;
+
+/**
+ * Yes I know.
+ */
+public interface EnvTypeable {
+    EnvType getType();
+}
